1_Publication/requirements-test.txt
# Testing dependencies for 1_Publication

# Core testing framework
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0

# Dependencies from main modules
requests>=2.31.0
google-generativeai>=0.3.0
pydantic>=2.0.0

# Additional testing utilities
faker>=20.0.0
freezegun>=1.4.0
responses>=0.24.0
